+++ NEWS.md
@@ -1,5 +1,6 @@
 # glebrt 0.6.0
 
+- Add tests for logical vectors: `some()` and `same()`
 - Set up testing
 
 # glebrt 0.5.0